requirements.txt
requests
flask
pywebview